split to library
