+++ Build1/zmachine.Library/ConsoleIO.cs
