+++ Build1/zmachine.Library/Lex.cs
