+++ Build1/zmachine.Library/Machine.OpcodeHandler.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
+﻿using System.Diagnostics;
 
 namespace zmachine
 {
