+++ Build1/zmachine.Library/Machine.State.cs
