+++ Build1/zmachine.Library/Machine.cs
