+++ Build1/zmachine.Library/Memory.cs
