+++ Build1/zmachine.Library/OperandType.cs
