+++ Build1/zmachine.Library/StaticIO.cs
