+++ Build1/zmachine.Library/crc32.cs
@@ -1,5 +1,4 @@
-﻿using System;
-public class Crc32
+﻿public class Crc32
 {
     uint[] table;
