+++ Build1/zmachine.Library/zmachine.Library.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <RootNamespace>zmachine.Library</RootNamespace>
+	<AssemblyName>zmachine.Library</AssemblyName>
+  </PropertyGroup>
+	<ItemGroup>
+		<PackageReference Include="Unity3D.SDK" Version="2021.1.14.1" />
+	</ItemGroup>
+</Project>
